+++ jablib/src/main/java/org/jabref/logic/citationkeypattern/CitationKeyGeneratorTestUtils.java
@@ -1,5 +1,7 @@
 package org.jabref.logic.citationkeypattern;
 
+import javafx.beans.property.SimpleObjectProperty;
+
 import static org.jabref.logic.citationkeypattern.CitationKeyGenerator.DEFAULT_UNWANTED_CHARACTERS;
 
 public class CitationKeyGeneratorTestUtils {
@@ -15,8 +17,7 @@ public class CitationKeyGeneratorTestUtils {
                 "",
                 DEFAULT_UNWANTED_CHARACTERS,
                 GlobalCitationKeyPatterns.fromPattern("[auth][year]"),
-                "",
-                ','
+                new SimpleObjectProperty<>(',')
         );
     }
 }